Moved all gobject enums in a single header
This is really for readability, I have no other justification. Also fixed minor
bug in the throttleing scheme: sometimes, when there was only 1 frame left to pull
from the cine, the code would get stuck in a loop because of a condition stopping
the code from pulling a single image. This has been changed to making sure the count
positive.
