+++ uca-phantom-properties.h
@@ -1,0 +1,419 @@
+#ifndef __UCA_PHANTOM_PROPERTIES_H
+#define __UCA_PHANTOM_PROPERTIES_H
+
+#include <glib.h>
+#include <uca/uca-camera.h>
+
+// For uca-phantom-communicate.h
+enum UcaPhantomCommunicateProperties {
+    PROP_COM_PHANTOM_IP = 1,
+    PROP_COM_PHANTOM_XIP,
+    PROP_COM_NETCARD_IP,
+    PROP_COM_NETCARD_XIP,
+    PROP_COM_NETCARD,
+    PROP_COM_XNETCARD,
+    PROP_COM_XENABLED,
+    PROP_COM_CONTROL_PORT,
+    PROP_COM_PHANTOM_IPSOURCE,
+    N_COM_PROPERTIES
+};
+
+// For uca-phantom-camera.h
+enum UcaPhantomCameraProperties {
+    PROP_NB_POST_TRIGGER_FRAMES = N_BASE_PROPERTIES,
+    PROP_NB_PRE_TRIGGER_FRAMES,
+    PROP_SYNC_MODE,
+    PROP_ACQUISITION_MODE,
+    PROP_IMAGE_FORMAT,
+    PROP_TIMESTAMP_FORMAT,
+    PROP_XNETCARD,
+    PROP_XENABLED,
+    PROP_SENSOR_PHYSICAL_HEIGHT,
+    PROP_SENSOR_PHYSICAL_WIDTH,
+    PROP_MAX_SENSOR_RESOLUTION_WIDTH,
+    PROP_MAX_SENSOR_RESOLUTION_HEIGHT,
+    PROP_NUM_CINES,
+    PROP_LIVEIMAGES,
+    N_PHANTOM_PROPERTIES
+};
+
+enum PhantomUnitIds {
+    UNIT_INFO_SENSOR,
+    UNIT_INFO_SNSVERSION,
+    UNIT_INFO_CFA,
+    UNIT_INFO_FILTER,
+    UNIT_INFO_HWVER,
+    UNIT_INFO_KERNEL,
+    UNIT_INFO_SWVER,
+    UNIT_INFO_XVER,
+    UNIT_INFO_MODEL,
+    UNIT_INFO_PVER,
+    UNIT_INFO_SVER,
+    UNIT_INFO_FVER,
+    UNIT_INFO_SERIAL,
+    UNIT_INFO_NAME,
+    UNIT_INFO_FEATURES,
+    UNIT_INFO_IMGFORMATS,
+    UNIT_INFO_VIDEOSYSTEMS,
+    UNIT_INFO_MAXCINES,
+    UNIT_INFO_XMAX,
+    UNIT_INFO_YMAX,
+    UNIT_INFO_XINC,
+    UNIT_INFO_YINC,
+    UNIT_INFO_WINX,
+    UNIT_INFO_WINY,
+    UNIT_INFO_KERNSZ,
+    UNIT_INFO_MEMSZ,
+    UNIT_INFO_CINEMEM,
+    UNIT_INFO_MDEPTHS,
+    UNIT_INFO_EXPDEAD,
+    UNIT_INFO_MINEXP,
+    UNIT_INFO_XBLOCK,
+    UNIT_INFO_YBLOCK,
+    UNIT_INFO_PIXPS,
+    UNIT_INFO_ROTPS,
+    UNIT_INFO_FOTPS,
+    UNIT_INFO_MINFRATE,
+    UNIT_INFO_MAXRATE,
+    UNIT_INFO_TMODEL,
+    UNIT_INFO_MAGTP,
+    UNIT_INFO_RTOBYTEPS,
+    UNIT_INFO_RTOPACKET,
+    UNIT_INFO_RTOPACKETOVHEAD,
+    UNIT_INFO_RTOFROVHEAD,
+    UNIT_INFO_RTOCHANNELS,
+    UNIT_INFO_BASECHROMA,
+    UNIT_INFO_BASEEI,
+    UNIT_INFO_SNSTEMP,
+    UNIT_INFO_TEPOWER,
+    UNIT_INFO_CAMTEMP,
+    UNIT_INFO_FANPOWER,
+    UNIT_INFO_BATTI,
+    UNIT_INFO_BATTV,
+    UNIT_INFO_BATTSTATE,
+    UNIT_INFO_BATTTIMER,
+    // UNIT_INFO_GENLOCKSTAT,
+    UNIT_INFO_MODES,
+    UNIT_PRESET_TONE1,
+    UNIT_PRESET_TONE2,
+    UNIT_PRESET_TONE3,
+    UNIT_PRESET_TONE4,
+    UNIT_PRESET_MATRIX1,
+    UNIT_PRESET_MATRIX2,
+    UNIT_PRESET_MATRIX3,
+    UNIT_PRESET_MATRIX4,
+    UNIT_PRESET_FILTER1,
+    UNIT_PRESET_FILTER2,
+    UNIT_PRESET_FILTER3,
+    UNIT_PRESET_FILTER4,
+    UNIT_PRESET_EI1,
+    UNIT_PRESET_EI2,
+    UNIT_PRESET_EI3,
+    UNIT_PRESET_EI4,
+    UNIT_PRESET_EI5,
+    UNIT_PRESET_EI6,
+    UNIT_PRESET_EI7,
+    UNIT_PRESET_EI8,
+    UNIT_CAM_SYNCIMG,
+    UNIT_CAM_FRDELAY,
+    UNIT_CAM_RTOEN,
+    UNIT_CAM_RTOTFR,
+    UNIT_CAM_MEMBPP,
+    UNIT_CAM_TRIGPOL,
+    UNIT_CAM_TRIGFILT,
+    UNIT_CAM_STARTONACQ,
+    UNIT_CAM_AUX1MODE,
+    UNIT_CAM_AUX2MODE,
+    UNIT_CAM_AUX3MODE,
+    UNIT_CAM_TSFORMAT,
+    UNIT_CAM_TCMODE,
+    UNIT_CAM_MASTER,
+    UNIT_CAM_APOFFDIS,
+    UNIT_CAM_LONGREADY,
+    UNIT_CAM_CINES,
+    UNIT_CAM_DARK,
+    UNIT_CAM_TSETSNS,
+    UNIT_CAM_TSETCAM,
+    // UNIT_CAM_TZ,
+    UNIT_CAM_MODE,
+    UNIT_AUTO_VIDEOPLAY,
+    UNIT_AUTO_FLASHSAVE,
+    UNIT_AUTO_FILESAVE,
+    UNIT_AUTO_ACQRESTART,
+    UNIT_AUTO_BREF,
+    UNIT_AUTO_FIRSTFRAME,
+    UNIT_AUTO_LASTFRAME,
+    UNIT_AUTO_LOOPS,
+    UNIT_AUTO_SPEED,
+    UNIT_AUTO_PROGRESS,
+    UNIT_AUTO_BREFPROGRESS,
+    UNIT_AUTO_TRIGGER_X,
+    UNIT_AUTO_TRIGGER_Y,
+    UNIT_AUTO_TRIGGER_W,
+    UNIT_AUTO_TRIGGER_H,
+    UNIT_AUTO_TRIGGER_THRESHOLD,
+    UNIT_AUTO_TRIGGER_AREA,
+    UNIT_AUTO_TRIGGER_SPEED,
+    UNIT_AUTO_TRIGGER_MODE,
+    UNIT_DEFC_RES,
+    UNIT_DEFC_RATE,
+    UNIT_DEFC_EXP,
+    UNIT_DEFC_EDREXP,
+    UNIT_DEFC_PTFRAMES,
+    UNIT_DEFC_SHOFF,
+    UNIT_DEFC_RAMP,
+    UNIT_DEFC_BCOUNT,
+    UNIT_DEFC_BPERIOD,
+    UNIT_DEFC_HQENABLE,
+    UNIT_DEFC_DECIMATION,
+    UNIT_DEFC_FRCOUNT,
+    UNIT_DEFC_FRSIZE,
+    UNIT_DEFC_AEXPMODE,
+    UNIT_DEFC_AEXPCOMP,
+    UNIT_DEFC_META_OX,
+    UNIT_DEFC_META_OY,
+    UNIT_DEFC_META_W,
+    UNIT_DEFC_META_H,
+    UNIT_DEFC_META_OW,
+    UNIT_DEFC_META_OH,
+    UNIT_DEFC_META_CROP,
+    // UNIT_DEFC_META_SCALE,
+    UNIT_IRIG_SEC,
+    UNIT_IRIG_YEARBEGIN,
+    UNIT_IRIG_FLAGS,
+    UNIT_IRIG_SIGNAL,
+    UNIT_IRIG_GPS,
+    // UNIT_IRIG_RANGE,
+    UNIT_META_NAME,
+    UNIT_META_LENS,
+    UNIT_META_FSTOP,
+    UNIT_META_FLEN,
+    UNIT_META_COMMENT,
+    UNIT_META_XSET,
+    UNIT_VIDEO_SYSTEM,
+    UNIT_VIDEO_OUTPUT,
+    UNIT_VIDEO_FIELDS,
+    UNIT_VIDEO_WIDESCREEN,
+    UNIT_VIDEO_GENLOCK,
+    UNIT_VIDEO_VFMODE,
+    UNIT_VIDEO_OSDDIS,
+    UNIT_VIDEO_OSDUT,
+    UNIT_VIDEO_PAX,
+    UNIT_VIDEO_PAY,
+    UNIT_VIDEO_PAOX,
+    UNIT_VIDEO_PAOY,
+    UNIT_VIDEO_UZOOM,
+    UNIT_VIDEO_VOX,
+    UNIT_VIDEO_VOY,
+    UNIT_VIDEO_VOW,
+    UNIT_VIDEO_VOH,
+    UNIT_VIDEO_VW,
+    UNIT_VIDEO_VH,
+    UNIT_VIDEO_ADJ_RED,
+    UNIT_VIDEO_ADJ_GREEN,
+    UNIT_VIDEO_ADJ_BLUE,
+    UNIT_VIDEO_ADJ_TOE,
+    UNIT_VIDEO_ADJ_GAMMA,
+    UNIT_VIDEO_ADJ_RGAMMA,
+    UNIT_VIDEO_ADJ_BGAMMA,
+    UNIT_VIDEO_ADJ_GAIN,
+    UNIT_VIDEO_ADJ_OFFSET,
+    UNIT_VIDEO_ADJ_FLARE,
+    UNIT_VIDEO_ADJ_HUE,
+    UNIT_VIDEO_ADJ_SAT,
+    UNIT_VIDEO_ADJ_RPED,
+    UNIT_VIDEO_ADJ_GPED,
+    UNIT_VIDEO_ADJ_BPED,
+    UNIT_VIDEO_ADJ_CHROMA,
+    UNIT_VIDEO_ADJ_TONE,
+    UNIT_VIDEO_ADJ_MATRIX,
+    UNIT_VIDEO_ADJ_LOG,
+    UNIT_VIDEO_ADJ_SDIMIN,
+    UNIT_VIDEO_ADJ_SDIMAX,
+    UNIT_VIDEO_ADJ_CMATRIX,
+    UNIT_VIDEO_ADJ_FILTER,
+    UNIT_VIDEO_ADJ_UMATRIX,
+    UNIT_VIDEO_ADJ_WBTEMP,
+    UNIT_VIDEO_ADJ_WBCC,
+    UNIT_VIDEO_ADJ_WBRED,
+    UNIT_VIDEO_ADJ_WBBlUE,
+    UNIT_VIDEO_ADJ_MMSAT,
+    UNIT_VIDEO_ADJ_MMHUE,
+    UNIT_VIDEO_ADJ_EDGEG,
+    UNIT_VIDEO_ADJ_EDGER,
+    UNIT_VIDEO_ADJ_EDGETH,
+    UNIT_VIDEO_PLAY_LIVE,
+    UNIT_VIDEO_PLAY_CINE,
+    UNIT_VIDEO_PLAY_MAG,
+    UNIT_VIDEO_PLAY_FN,
+    UNIT_VIDEO_PLAY_IN,
+    UNIT_VIDEO_PLAY_OUT,
+    UNIT_VIDEO_PLAY_SPEED,
+    UNIT_VIDEO_PLAY_STEP,
+    UNIT_VIDEO_PLAY_FRCOUNT,
+    UNIT_ETH_IP,
+    UNIT_ETH_NETMASK,
+    UNIT_ETH_BROADCAST,
+    UNIT_ETH_GATEWAY,
+    UNIT_ETH_MTU,
+    UNIT_ETH_XIP,
+    UNIT_ETH_XNETMASK,
+    UNIT_ETH_XBROADCAST,
+    UNIT_MAG_STATE,
+    UNIT_MAG_PROGRESS,
+    UNIT_MAG_PROTECT,
+    UNIT_MAG_SIZE,
+    UNIT_MAG_USED,
+    UNIT_MAG_TAKES,
+    UNIT_MAG_VERSION,
+    UNIT_MAG_ID,
+    UNIT_MAG_RUNSTOP,
+    UNIT_MAG_TYPE,
+    UNIT_CF_STATE,
+    UNIT_CF_ACTION,
+    UNIT_CF_SIZE,
+    UNIT_CF_USED,
+    UNIT_CF_PROGRESS,
+    UNIT_CF_ERRCODE,
+    UNIT_USETS_NSETS,
+    UNIT_USETS_VALID_NUMBER,
+    UNIT_C_STATE,
+    UNIT_C_FRCOUNT,
+    UNIT_C_FIRSTFR,
+    UNIT_C_LASTFR,
+    UNIT_C_FORMAT,
+    UNIT_C_IN,
+    UNIT_C_OUT,
+    UNIT_C_START,
+    UNIT_C_LEN,
+    UNIT_C_FRSIZE,
+    UNIT_C_FRSPACE,
+    UNIT_C_TRIGTIME_SECS,
+    UNIT_C_TRIGTIME_FRAC,
+    UNIT_C_META_PBRATE,
+    UNIT_C_META_TCRATE,
+    UNIT_C_META_UUID,
+    UNIT_C_META_SYSTEM,
+    UNIT_C_META_TRIGTC,
+    UNIT_C_META_PAX,
+    UNIT_C_META_PAY,
+    UNIT_C_META_PAOX,
+    UNIT_C_META_PAOY,
+    UNIT_C_META_VOX,
+    UNIT_C_META_VOY,
+    UNIT_C_META_VOW,
+    UNIT_C_META_VOH,
+    UNIT_C_META_VW,
+    UNIT_C_META_VH,
+    UNIT_C_META_OX,
+    UNIT_C_META_OY,
+    UNIT_C_META_OW,
+    UNIT_C_META_OH,
+    UNIT_C_META_W,
+    UNIT_C_META_H,
+    UNIT_C_META_CROP,
+    UNIT_C_META_RESIZE,
+    UNIT_C_META_GPS,
+    N_CINE_PROPERTIES,
+    // UNIT_HW_VPIX,
+    // UNIT_HW_VPIXL,
+    // UNIT_HW_VMEMH,
+    // UNIT_HW_VMEML,
+    // UNIT_HW_VRESH,
+    // UNIT_HW_VRESL,
+    // UNIT_HW_VRESDS,
+    // UNIT_HW_DCBLACK,
+    // UNIT_HW_BBIAS,
+    // UNIT_HW_TBIAS,
+    // UNIT_HW_TRBIAS,
+    // UNIT_HW_BRBIAS,
+    // UNIT_HW_VMLPRED,
+    // UNIT_HW_PREPW,
+    // UNIT_HW_PRESMPD,
+    // UNIT_HW_SMPPW,
+    // UNIT_HW_SMPVMHD,
+    // UNIT_HW_VMHRESD,
+    // UNIT_HW_RESPW,
+    // UNIT_HW_RESWAIT,
+    // UNIT_HW_EXPADJ,
+    // UNIT_HW_SEL2W,
+    // UNIT_HW_SHKOLMODE,
+    // UNIT_HW_RESLONG,
+    // UNIT_HW_SNSFLAGS,
+    // UNIT_HW_NSFLOAD,
+    // UNIT_HW_COLLOAD,
+    // UNIT_HW_OUTLOAD,
+    // UNIT_HW_FREQTRIM,
+    // UNIT_HW_IRIGPHASE,
+    // UNIT_HW_IRIGMODPHASE,
+    // UNIT_HW_MCODE,
+    // UNIT_HW_MEMPHASE,
+    // UNIT_HW_MEMFLAGS,
+    // UNIT_HW_ADPHASE,
+    // UNIT_HW_ADCFLAGS,
+    // UNIT_HW_DCPHASE,
+    // UNIT_HW_TOUCHDX,
+    // UNIT_HW_TOUCHDY,
+    // UNIT_HW_TOUCHSX,
+    // UNIT_HW_TOUCHSY,
+    // UNIT_HW_IGAIN,
+    // UNIT_HW_COLORCAL,
+    // UNIT_HW_SNTEMPCAL,
+};
+
+enum {
+    N_UNIT_PROPERTIES = UNIT_USETS_VALID_NUMBER,
+};
+
+enum PhantomCommandIds{
+    CMD_GET, //
+    CMD_SET, //
+    CMD_START_RECORDING_IN_A_CINE, //
+    CMD_DELETE_A_CINE, //
+    CMD_RELEASE_A_CINE, //
+    CMD_SOFTWARE_TRIGGER, //
+    CMD_GET_CINE_STATES, //
+    CMD_START_DATA_CONNECTION, //
+    CMD_ATTACH, //
+    CMD_GET_IMAGES, //
+    CMD_GET_XIMAGES, //
+    CMD_GET_TIMESTAMPS, //
+    CMD_SHOW_NETWORK_INTERFACE_CONFIGURATION, //
+    CMD_SHOW_ROUTING_TABLE, //
+    CMD_PARTITION_CINE_MEMORY, //
+    CMD_PERFORM_WHITE_BALANCE, //
+    CMD_PERFORM_BLACK_REFERENCE, //
+    CMD_PRNU_CORRECTION_UPDATE, //
+    CMD_BLACK_REFERENCE_UPDATE, //
+    CMD_FLASH_ERASE, //
+    CMD_FLASH_SAVE, //
+    CMD_STORAGE_DEVICE_SAVE, //
+    CMD_RETRIEVE_INTERNAL_CAMERA_LOG, //
+    CMD_PLAY_FRAMES_ON_VIDEO_OUTPUT, //
+    CMD_CHECK_FOR_VARIABLE_CHANGES, //
+    CMD_ENABLE_STATUS_CHANGE_NOTIFICATIONS, //
+    CMD_SAVE_FACTORY_DEFAULTS, //
+    CMD_LOAD_FACTORY_DEFAULTS, //
+    CMD_SAVE_USER_SETTINGS, //
+    CMD_LOAD_USER_SETTINGS, //
+    CMD_ERASE_USER_SETTINGS, //
+    CMD_LIST_USER_SETTINGS, //
+    CMD_DEBUG_CONSOLE_MODE,//
+    CMD_SET_LENS_APERTURE,//
+    CMD_MOVE_FOCUS,//
+    CMD_ISSUE_LENS_MOUNT_COMMAND,//
+    CMD_CHANGE_SERIAL_LINE_BAUDE_RATE,//
+    CMD_CAMERA_CALIBRATION,//
+    CMD_SYSTEM_MONITOR,//
+    CMD_GENERATE_TEST_IMAGE,//
+    CMD_SET_REAL_TIME_CLOCK,//
+    CMD_LIST_FILES_ON_STORAGE_DEVICE,//
+    CMD_REMOVE_FILE_FROM_STORAGE_DEVICE,//
+    CMD_FORMATE_STORAGE_DEVICE,//
+    CMD_FILE_READ_DATA_FROM_STORAGE_DEVICE,//
+    CMD_USE_COLOR_PRESET,//
+    CMD_SET_MULTI_MATRIX_AXIS,//
+    N_UNIT_COMMANDS
+};
+
+#endif
