+++ uca-phantom-variables.c
@@ -47,7 +47,7 @@ PhantomUnit variables[] = {
     {"info.rtopacket", G_TYPE_UINT, G_PARAM_READABLE, UNIT_INFO_RTOPACKET, "Size of a packet in bytes."},
     {"info.rtopacketovhead", G_TYPE_UINT, G_PARAM_READABLE, UNIT_INFO_RTOPACKETOVHEAD, "Overhead per packet in bytes."},
     {"info.rtofrovhead", G_TYPE_UINT, G_PARAM_READABLE, UNIT_INFO_RTOFROVHEAD, "Overhead per frame in bytes."},
-    {"info.rtochannels", G_TYPE_UINT, G_PARAM_READABLE, UNIT_INFO_RTOCHANNELS, "Number of RTO channels."},
+    {"info.rto_channels", G_TYPE_UINT, G_PARAM_READABLE, UNIT_INFO_RTOCHANNELS, "Number of RTO channels."},
     {"info.basechroma", G_TYPE_FLOAT, G_PARAM_READABLE, UNIT_INFO_BASECHROMA, "Default chroma gain used when the color matrix is off."},
     {"info.baseei", G_TYPE_UINT, G_PARAM_READABLE, UNIT_INFO_BASEEI, "Base Exposure Index, in ISO units. Exposing for this index should result in a properly exposed image with the camera at minimum video gain, a neutral tone curve and a gamma of 2.2."},
     {"info.snstemp", G_TYPE_INT, G_PARAM_READABLE, UNIT_INFO_SNSTEMP, "Sensor die temperature."},
@@ -58,7 +58,7 @@ PhantomUnit variables[] = {
     {"info.battv", G_TYPE_UINT, G_PARAM_READABLE, UNIT_INFO_BATTV, "Battery voltage in millivolts."},
     {"info.battstate", G_TYPE_UINT, G_PARAM_READABLE, UNIT_INFO_BATTSTATE, "Battery status: 0=no battery, 1=charging, 2=full, 3 =discharging, 4=low."},
     {"info.batttimer", G_TYPE_UINT, G_PARAM_READABLE, UNIT_INFO_BATTTIMER, "Number of seconds until camera will turn down by itself, 0=stay on until battery empty."},
-    {"info.genlockstat", G_TYPE_UINT, G_PARAM_READABLE, UNIT_INFO_GENLOCKSTAT, "State of genlock system: 0=poweroff, 1=idle, 2=measure line length, 3=wait for line lock, 4=measure field length, 5=enter field lock, 6=force field lock, 7=field lock check, 8=locked, 9=bad signal, 10=locked and time code present."},
+    // {"info.genlockstat", G_TYPE_UINT, G_PARAM_READABLE, UNIT_INFO_GENLOCKSTAT, "State of genlock system: 0=poweroff, 1=idle, 2=measure line length, 3=wait for line lock, 4=measure field length, 5=enter field lock, 6=force field lock, 7=field lock check, 8=locked, 9=bad signal, 10=locked and time code present."},
     {"info.modes", G_TYPE_STRING, G_PARAM_READABLE, UNIT_INFO_MODES, "List of supported modes. The modes are assigned sequential numbers starting from zero, which correspond to the labels in this list. If a mode number is not used, the corresponding label is \u201cNA\u201d."},
     {"preset.tone1", G_TYPE_STRING, G_PARAM_READWRITE, UNIT_PRESET_TONE1, "Storage for tone curve presets."},
     {"preset.tone2", G_TYPE_STRING, G_PARAM_READWRITE, UNIT_PRESET_TONE2, "Storage for tone curve presets."},
@@ -100,7 +100,7 @@ PhantomUnit variables[] = {
     {"cam.dark", G_TYPE_INT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_CAM_DARK, "Close (1) / Open (0) the mechanical shutter."},
     {"cam.tsetsns", G_TYPE_UINT, G_PARAM_READABLE, UNIT_CAM_TSETSNS, "Sensor temperature set point."},
     {"cam.tsetcam", G_TYPE_UINT, G_PARAM_READABLE, UNIT_CAM_TSETCAM, "Camera temperature set point."},
-    {"cam.tz", G_TYPE_UINT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_CAM_TZ, "Difference in seconds between local time and UTC. Used when displaying local time on the OSD. All time values in the protocol are UTC."},
+    // {"cam.tz", G_TYPE_UINT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_CAM_TZ, "Difference in seconds between local time and UTC. Used when displaying local time on the OSD. All time values in the protocol are UTC."},
     {"cam.mode", G_TYPE_UINT, G_PARAM_READABLE, UNIT_CAM_MODE, "Current camera mode number. The corresponding mode name can be found using info.modes. Use iload to change the mode."},
     {"auto.videoplay", G_TYPE_UINT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_AUTO_VIDEOPLAY, "When non-zero, the camera will play back the recorded cine on video immediately after the recording ends and the autosave operations are finished. A rec command or the pretrigger signal will abort a video play operation."},
     {"auto.flashsave", G_TYPE_UINT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_AUTO_FLASHSAVE, "When non-zero, a camera will save the just-recorded cine to a cinemag. If the cinemag fills up (or is not installed), this flag is cleared."},
@@ -143,13 +143,13 @@ PhantomUnit variables[] = {
     {"defc.meta.ow", G_TYPE_INT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_DEFC_META_OW, "Window width after software crop and scale in pixels."},
     {"defc.meta.oh", G_TYPE_INT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_DEFC_META_OH, "Window height after software crop and scale in pixels."},
     {"defc.meta.crop", G_TYPE_INT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_DEFC_META_CROP, "Enable software crop."},
-    {"defc.meta.scale", G_TYPE_INT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_DEFC_META_SCALE, "Enable software scale."},
+    // {"defc.meta.scale", G_TYPE_INT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_DEFC_META_SCALE, "Enable software scale."},
     {"irig.sec", G_TYPE_UINT, G_PARAM_READABLE, UNIT_IRIG_SEC, "Current time of the camera timebase in seconds since 1970."},
     {"irig.yearbegin", G_TYPE_UINT, G_PARAM_READABLE, UNIT_IRIG_YEARBEGIN, "Number of seconds since 1970 to the beginning of the current year; used to add year information to the irig time stamps."},
     {"irig.flags", G_TYPE_STRING, G_PARAM_READABLE, UNIT_IRIG_FLAGS, "Status flags for the timebase unit; see below."},
     {"irig.signal", G_TYPE_STRING, G_PARAM_READABLE, UNIT_IRIG_SIGNAL, "Textual description of the incoming time code signal."},
     {"irig.gps", G_TYPE_STRING, G_PARAM_READABLE, UNIT_IRIG_GPS, "Current position and status info as received from a GPS receiver. It contains the GPRMC, GPGGA and GPGSA messages"},
-    {"irig.range", G_TYPE_STRING, G_PARAM_READABLE, UNIT_IRIG_RANGE, "String representation of the current range data being received."},
+    // {"irig.range", G_TYPE_STRING, G_PARAM_READABLE, UNIT_IRIG_RANGE, "String representation of the current range data being received."},
     {"meta.name", G_TYPE_STRING, G_PARAM_READWRITE, UNIT_META_NAME, "A 256 bytes string used to name the recorded cines. Succesive names are automatically generated by appending an index to meta.name. If meta.name already contains a trailing numeric value, this value is used as a start value for the index."},
     {"meta.lens", G_TYPE_STRING, G_PARAM_READABLE, UNIT_META_LENS, "focal length range and aperture range of a supported \u201cintelligent\u201d lens. When no active lens mount is fitted, or no lens is mounted, this field is set to a null string."},
     {"meta.fstop", G_TYPE_FLOAT, G_PARAM_READABLE, UNIT_META_FSTOP, "current fstop value of the fitted lens."},
@@ -163,7 +163,7 @@ PhantomUnit variables[] = {
     {"video.genlock", G_TYPE_UINT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_VIDEO_GENLOCK, "1 = Enable genlock. 3= Enable genlock and reading the VITC (linear time code) from the video return input."},
     {"video.vfmode", G_TYPE_UINT, G_PARAM_READWRITE, UNIT_VIDEO_VFMODE, "Viewfinder mode. 0=normal, 1=threshold, 2=zoom, 4=zoom more, 8=show return signal (on some cameras)."},
     {"video.osddis", G_TYPE_INT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_VIDEO_OSDDIS, "Turns off the OSD on selected outputs. 1: OSD off on analog and HSDSI1; 2: OSD off on HDSDI2; 3: OSD off on all outputs."},
-    {"video.osdut", G_TYPE_BOOLEAN, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_VIDEO_OSDUT, "Show UT timestamps instead of local time on the OSD."},
+    {"video.osd_ut", G_TYPE_BOOLEAN, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_VIDEO_OSDUT, "Show UT timestamps instead of local time on the OSD."},
     {"video.pax", G_TYPE_UINT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_VIDEO_PAX, "Production area X size."},
     {"video.pay", G_TYPE_UINT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_VIDEO_PAY, "Production area Y size."},
     {"video.paox", G_TYPE_INT, G_PARAM_READWRITE | G_PARAM_CONSTRUCT, UNIT_VIDEO_PAOX, "Production area X offset from center of picture."},
